Combine testsuite.go files into the associated test.go files
Combine testsuite.go files into the associated test.go files
