+++ internal/pkg/impl/common/container/progress/progress_action_test.go
@@ -9,3 +9,17 @@ import (
 func TestComposeProgressActionTestSuite(t *testing.T) {
 	suite.Run(t, new(ComposeProgressActionTestSuite))
 }
+
+type ComposeProgressActionTestSuite struct {
+	suite.Suite
+}
+
+func (t *ComposeProgressActionTestSuite) TestString() {
+	t.Equal("Building", Build.String())
+	t.Equal("Creating", Create.String())
+	t.Equal("Starting", Start.String())
+	t.Equal("Stopping", Stop.String())
+	t.Equal("Removing", Remove.String())
+
+	t.Equal("Unknown", Unknown.String())
+}
