--- internal/pkg/impl/common/logging/blackhole_handler_testsuite.go
@@ -1,29 +1,0 @@
-package logging
-
-import (
-	"context"
-	"log/slog"
-
-	"github.com/stretchr/testify/suite"
-)
-
-type BlackholeHandlerTestSuite struct {
-	suite.Suite
-}
-
-func (t *BlackholeHandlerTestSuite) TestHandler() {
-	handler := NewBlackHoleHandler()
-
-	t.False(handler.Enabled(context.Background(), slog.LevelDebug))
-
-	t.Nil(handler.Handle(context.Background(), slog.Record{
-		Level:   slog.LevelDebug,
-		Message: "This is a test message",
-	}))
-
-	t.Equal(handler, handler.WithAttrs([]slog.Attr{
-		slog.String("key", "value"),
-	}))
-
-	t.Equal(handler, handler.WithGroup("test-group"))
-}
