--- internal/pkg/impl/common/logging/log_handler_builder_testsuite.go
@@ -1,43 +1,0 @@
-package logging
-
-import (
-	"log/slog"
-
-	"github.com/stretchr/testify/suite"
-)
-
-type LogHandlerBuilderTestSuite struct {
-	suite.Suite
-
-	logFilePath string
-	builder     *LogHandlerBuilder
-}
-
-func (t *LogHandlerBuilderTestSuite) SetupTest() {
-	t.builder = NewLogHandlerBuilder()
-	t.logFilePath = t.T().TempDir() + "/logfile.log"
-}
-
-func (t *LogHandlerBuilderTestSuite) TestTextHandler() {
-	t.builder.WithLogFilePath(t.logFilePath).
-		WithLogLevel("error").
-		WithLogHandlerName("text")
-
-	handler, err := t.builder.Build()
-	t.NoError(err)
-	t.NotNil(handler)
-
-	t.IsType(&slog.TextHandler{}, handler)
-}
-
-func (t *LogHandlerBuilderTestSuite) TestJsonHandler() {
-	t.builder.WithLogFilePath(t.logFilePath).
-		WithLogLevel("error").
-		WithLogHandlerName("json")
-
-	handler, err := t.builder.Build()
-	t.NoError(err)
-	t.NotNil(handler)
-
-	t.IsType(&slog.JSONHandler{}, handler)
-}
