+++ internal/pkg/impl/host/project/service_workflows_config_test.go
@@ -9,3 +9,12 @@ import (
 func TestServiceWorkflowsTestSuite(t *testing.T) {
 	suite.Run(t, new(ServiceWorkflowsTestSuite))
 }
+
+type ServiceWorkflowsTestSuite struct {
+	suite.Suite
+}
+
+func (t *ServiceWorkflowsTestSuite) TestNewServiceWorkflows() {
+	serviceWorkflow := NewServiceWorkflows()
+	t.NotNil(serviceWorkflow)
+}
