--- internal/pkg/impl/host/project/service_workflows_config_testsuite.go
@@ -1,12 +1,0 @@
-package project
-
-import "github.com/stretchr/testify/suite"
-
-type ServiceWorkflowsTestSuite struct {
-	suite.Suite
-}
-
-func (t *ServiceWorkflowsTestSuite) TestNewServiceWorkflows() {
-	serviceWorkflow := NewServiceWorkflows()
-	t.NotNil(serviceWorkflow)
-}
